+++ lmdeploy/pytorch_poc/engine/engine.py
@@ -1,36 +1,27 @@
 # Copyright (c) OpenMMLab. All rights reserved.
 import enum
 import itertools
-import json
-import os
 import time
 from dataclasses import dataclass
 from queue import Queue
-from threading import Thread
-from typing import Any, Callable, Dict, List
+from threading import Lock, Thread
+from typing import Any, Dict, List
 
 import torch
-import torch.distributed as dist
-from torch import multiprocessing as mp
-from torch.distributed._tensor import DeviceMesh, Replicate, distribute_tensor
-from transformers import AutoConfig, AutoModelForCausalLM
+from transformers import AutoConfig
 from transformers.generation.logits_process import (LogitsProcessorList,
                                                     TemperatureLogitsWarper,
                                                     TopKLogitsWarper,
                                                     TopPLogitsWarper)
-from transformers.utils import WEIGHTS_INDEX_NAME, WEIGHTS_NAME, cached_file
 
-from lmdeploy.pytorch.accel import LoadNoInit
 from lmdeploy.pytorch_poc.config import (CacheConfig, ModelConfig,
                                          SchedulerConfig)
+from lmdeploy.pytorch_poc.engine.model_agent import (BaseModelAgent,
+                                                     TPModelAgent)
 from lmdeploy.pytorch_poc.messages import (MessageStatus, SamplingParam,
                                            SchedulerSequence, SchedulerSession)
-from lmdeploy.pytorch_poc.models import patch
 from lmdeploy.pytorch_poc.paging import Scheduler
-from lmdeploy.pytorch_poc.utils import get_gpu_memory
 from lmdeploy.utils import get_logger
-
-from .cache_engine import CacheEngine
 
 logger = get_logger('lmdeploy')
 
@@ -203,29 +194,6 @@ class ModelContext:
                 v_state = v_state[token_num:]
 
 
-def _update_cache_config(model_config: ModelConfig,
-                         cache_config: CacheConfig,
-                         gpu_id: int = 0):
-    """Update the gpu mem and cpu mem according to model info.
-
-    Args:
-        model_config (ModelConfig): The config of the model.
-        cache_config (CacheConfig): The config of the cache info.
-        gpu_id (int): The GPU id to use.
-    """
-    GPU_MEM_PERCENT = 0.7
-    SWAP_SPACE = 4 * (1 << 30)
-    reserved_mem = torch.cuda.memory_reserved(gpu_id)
-    gpu_mem = (get_gpu_memory(gpu_id) - reserved_mem) * GPU_MEM_PERCENT
-    cpu_mem = SWAP_SPACE
-    cache_block_size = CacheEngine.get_cache_block_size(
-        cache_config.block_size, model_config)
-    if cache_config.num_cpu_blocks == 0:
-        cache_config.num_cpu_blocks = int(cpu_mem / cache_block_size)
-    if cache_config.num_gpu_blocks == 0:
-        cache_config.num_gpu_blocks = int(gpu_mem / cache_block_size)
-
-
 def _get_torch_dtype(config: Any, default: str = 'float16'):
     """Get the torch dtype from the model config.
 
@@ -235,215 +203,6 @@ def _get_torch_dtype(config: Any, default: str = 'float16'):
     """
     torch_dtype = getattr(config, 'torch_dtype', default)
     return eval(f'torch.{torch_dtype}')
-
-
-def _tp_model_loop(
-    rank: int,
-    model_path: str,
-    extra_args: List[str],
-    model_config: ModelConfig,
-    cache_config: CacheConfig,
-    in_que: mp.Queue,
-    out_que: mp.Queue,
-    world_size: int,
-    trust_remote_code=True,
-):
-    """Start model loops for tensor parallel model inference.
-
-    Args:
-        rank (int): Distribution rank.
-        model_path (int): Path of the hugging face model. Could be
-            local or online.
-        extra_args (List[str]): The extra arguments to add to the
-            patched model.
-        model_config (ModelConfig): The config of the model.
-        cache_config (CacheConfig): The config of the cache.
-        in_que (mp.Queue): Input queue. Used to receive model input.
-        out_que (mp.Queue): Output queue. Used to send the model output.
-        world_size (int): The distribution world size.
-    """
-    from accelerate import init_empty_weights
-
-    device_mesh = DeviceMesh('cuda', list(range(world_size)))
-
-    error_code = 0
-    error_type = None
-
-    try:
-        config = AutoConfig.from_pretrained(
-            model_path, trust_remote_code=trust_remote_code)
-        torch_dtype = _get_torch_dtype(config)
-        with init_empty_weights():
-            model = AutoModelForCausalLM.from_config(
-                config,
-                torch_dtype=torch_dtype,
-                trust_remote_code=trust_remote_code)
-
-        try:
-            torch_model_json_path = cached_file(model_path, WEIGHTS_INDEX_NAME)
-            with open(torch_model_json_path, mode='r') as f:
-                torch_model_json = json.load(f)
-
-            weight_map = torch_model_json['weight_map']
-
-            checkpoints = list(set(weight_map.values()))
-            checkpoints = [
-                cached_file(model_path, ckpt) for ckpt in checkpoints
-            ]
-        except Exception:
-            logger.warning(f'load failed, try load from {WEIGHTS_NAME}.')
-            checkpoints = [cached_file(model_path, WEIGHTS_NAME)]
-        patched_model = patch(
-            model,
-            extra_args=extra_args,
-            rank=rank,
-            world_size=world_size,
-            checkpoints=checkpoints,
-        )
-
-        _update_cache_config(model_config, cache_config)
-        cache_engine = CacheEngine(cache_config,
-                                   model_config,
-                                   rank=rank,
-                                   world_size=world_size)
-    except Exception as e:
-        error_code = 1
-        error_type = e
-
-    error_code = torch.tensor(error_code).cuda(rank)
-
-    dist.all_reduce(error_code)
-
-    error_code = error_code.item()
-    if error_code > 0:
-        all_errors = [None] * world_size
-        dist.all_gather_object(all_errors, error_type)
-        if rank == 0:
-            out_que.put((1, all_errors, cache_config))
-        return
-    else:
-        if rank == 0:
-            out_que.put((0, None, cache_config))
-
-    while True:
-        input_tensor_names = [
-            'input_ids',
-            'seq_length',
-            'attention_mask',
-            'position_ids',
-            'q_start_loc',
-        ]
-
-        if rank == 0:
-            inputs, swap_in_map, swap_out_map = in_que.get()
-            input_tensors = dict((name, t) for name, t in inputs.items()
-                                 if name in input_tensor_names)
-            input_metas = dict((name, (t.shape, t.dtype))
-                               for name, t in input_tensors.items())
-            input_metas['block_offsets'] = inputs['block_offsets']
-            input_metas['history_lengths'] = inputs['history_lengths']
-
-            objs = [input_metas, swap_in_map, swap_out_map]
-        else:
-            objs = [None, None, None]
-
-        # broadcast input shapes and dtypes
-        dist.broadcast_object_list(objs)
-
-        if rank != 0:
-            input_metas = objs[0]
-            input_tensors = dict((name, torch.empty(meta[0], dtype=meta[1]))
-                                 for name, meta in input_metas.items()
-                                 if name in input_tensor_names)
-
-        updated_inputs = dict()
-        for name, t in input_tensors.items():
-            updated_inputs[name] = distribute_tensor(t,
-                                                     device_mesh=device_mesh,
-                                                     placements=[Replicate()
-                                                                 ]).to_local()
-
-        inputs = updated_inputs
-        inputs['block_offsets'] = input_metas['block_offsets']
-        inputs['history_lengths'] = input_metas['history_lengths']
-
-        swap_in_map = objs[1]
-        swap_out_map = objs[2]
-
-        # swap in/out
-        issued_cache_op = False
-        if len(swap_in_map) > 0:
-            cache_engine.swap_in(swap_in_map)
-            issued_cache_op = True
-        if len(swap_out_map) > 0:
-            cache_engine.swap_out(swap_out_map)
-            issued_cache_op = True
-
-        if issued_cache_op:
-            cache_events = cache_engine.events
-        else:
-            cache_events = None
-
-        if cache_events is not None:
-            for event in cache_events:
-                event.wait()
-
-        with torch.no_grad():
-            output = patched_model(
-                input_ids=inputs['input_ids'],
-                position_ids=inputs['position_ids'],
-                attention_mask=inputs['attention_mask'],
-                past_key_values=cache_engine.gpu_cache,
-                return_dict=True,
-                output_attentions=False,
-                output_hidden_states=False,
-                use_origin=False,
-                context=ModelContext(
-                    block_offsets=inputs['block_offsets'],
-                    history_lengths=inputs['history_lengths'],
-                    position_ids=inputs['position_ids'],
-                    q_start_loc=inputs['q_start_loc'],
-                    seq_length=inputs['seq_length'],
-                    world_size=world_size,
-                ),
-                q_seq_info=(inputs['q_start_loc'], inputs['seq_length']),
-            )
-
-            if rank == 0:
-                out_que.put((0, output['logits'].cpu()))
-
-
-def _start_tp_process(rank: int,
-                      world_size: int,
-                      func: Callable,
-                      args: List = None,
-                      kwargs: Dict = None):
-    """Start the tensor parallel process.
-
-    Args:
-        rank (int): The distribution rank.
-        world_size (int): The distribution world size.
-        func (Callable): The function to be called in the process.
-        args (List): The arguments of the func.
-        kwargs (Dict): The keyword arguments of the func.
-    """
-    try:
-        os.environ['MASTER_ADDR'] = '127.0.0.1'
-        os.environ['MASTER_PORT'] = '29500'
-        dist.init_process_group('nccl', rank=rank, world_size=world_size)
-
-        with torch.cuda.device(rank), torch.no_grad():
-            if args is None:
-                args = tuple()
-            if kwargs is None:
-                kwargs = dict()
-            func(rank, *args, **kwargs)
-    except Exception as e:
-        from traceback import print_exc
-
-        logger.error(f'rank[{rank}]: {e}')
-        print_exc()
-        raise e
 
 
 class Engine:
@@ -522,102 +281,35 @@ class Engine:
                 dtype=torch_dtype,
             )
 
-        self.scheduler_config = scheduler_config
-        self.cache_config = cache_config
-        self.model_config = model_config
-        self.session_len = scheduler_config.max_session_len
-
         if tp == 1:
-            with LoadNoInit():
-                hf_model = AutoModelForCausalLM.from_pretrained(
-                    model_path,
-                    torch_dtype=torch_dtype,
-                    trust_remote_code=trust_remote_code)
-                hf_model.eval()
-
-            self.patched_model = patch(
-                hf_model, ['context', 'use_origin', 'q_seq_info']).cuda()
-            _update_cache_config(model_config, cache_config)
-
-            self.cache_engine = CacheEngine(cache_config, model_config)
-            logger.debug(
-                f'Initialize cache engine with {cache_config.num_gpu_blocks}'
-                f' gpu blocks and {cache_config.num_cpu_blocks} cpu blocks.')
-        else:
-            mp.set_start_method('spawn')
-            self.tp_model_in_que = mp.Queue(5)
-            self.tp_model_out_que = mp.Queue(5)
-
-            self.patch_model_tp(
+            self.model_agent = BaseModelAgent(
                 model_path,
-                ['context', 'use_origin', 'q_seq_info'],
                 model_config=model_config,
                 cache_config=cache_config,
-                in_que=self.tp_model_in_que,
-                out_que=self.tp_model_out_que,
+                trust_remote_code=trust_remote_code)
+        else:
+            self.model_agent = TPModelAgent(
+                model_path,
+                model_config=model_config,
+                cache_config=cache_config,
                 world_size=tp,
-            )
-
-            ret_code, e, cache_config = self.tp_model_out_que.get()
-            if ret_code != 0:
-                logger.error(f'Init tp model failed with error: {e}')
-                for err in e:
-                    if err is not None:
-                        raise err
-
-            # # have to update cache on host to support scheduler
-            # _update_cache_config(model_config, cache_config)
+                trust_remote_code=trust_remote_code)
+            cache_config = self.model_agent.cache_config
 
         self.scheduler = Scheduler(scheduler_config, cache_config)
 
         self.requests = Queue(scheduler_config.max_batches)
 
+        self.scheduler_config = scheduler_config
+        self.cache_config = cache_config
+        self.model_config = model_config
+        self.session_len = scheduler_config.max_session_len
+
         # create main thread
+        self.mutex = Lock()
         loop_threads = Thread(target=self.loop, daemon=True)
         loop_threads.start()
         self.loop_threads = loop_threads
-
-    def patch_model_tp(
-        self,
-        model_path: str,
-        extra_args: List[str],
-        model_config: ModelConfig,
-        cache_config: CacheConfig,
-        in_que: mp.Queue,
-        out_que: mp.Queue,
-        world_size: int,
-    ):
-        """Start tensor parallel sub process.
-
-        Args:
-            model_path (int): Path of the hugging face model.
-                Could be local or online.
-            extra_args (List[str]): The extra arguments to add to the
-                patched model.
-            model_config (ModelConfig): The config of the model.
-            cache_config (CacheConfig): The config of the cache.
-            in_que (mp.Queue): Input queue. Used to receive model input.
-            out_que (mp.Queue): Output queue. Used to send the model output.
-            world_size (int): The distribution world size.
-        """
-        self.mp_context = mp.spawn(
-            _start_tp_process,
-            args=(
-                world_size,
-                _tp_model_loop,
-                (model_path, extra_args),
-                dict(
-                    model_config=model_config,
-                    cache_config=cache_config,
-                    in_que=in_que,
-                    out_que=out_que,
-                    world_size=world_size,
-                ),
-            ),
-            nprocs=world_size,
-            join=False,
-            daemon=True,
-        )
 
     def create_instance(self, cuda_stream_id=0):
         """Create a turbomind instance.
@@ -736,57 +428,9 @@ class Engine:
             swap_in_map (Dict[int, int]): Cache maps to swap in.
             swap_out_map (Dict[int, int]): Cache maps to swap out.
         """
-        if self.tp == 1:
-            # swap in/out
-            issued_cache_op = False
-            if len(swap_in_map) > 0:
-                self.cache_engine.swap_in(swap_in_map)
-                issued_cache_op = True
-            if len(swap_out_map) > 0:
-                self.cache_engine.swap_out(swap_out_map)
-                issued_cache_op = True
-
-            if issued_cache_op:
-                cache_events = self.cache_engine.events
-            else:
-                cache_events = None
-
-            if cache_events is not None:
-                for event in cache_events:
-                    event.wait()
-
-            with torch.no_grad():
-                # forward
-                output = self.patched_model(
-                    input_ids=inputs['input_ids'],
-                    position_ids=inputs['position_ids'],
-                    attention_mask=inputs['attention_mask'],
-                    past_key_values=self.cache_engine.gpu_cache,
-                    return_dict=True,
-                    output_attentions=False,
-                    output_hidden_states=False,
-                    use_origin=False,
-                    context=ModelContext(
-                        block_offsets=inputs['block_offsets'],
-                        history_lengths=inputs['history_lengths'],
-                        position_ids=inputs['position_ids'],
-                        q_start_loc=inputs['q_start_loc'],
-                        seq_length=inputs['seq_length'],
-                    ),
-                    q_seq_info=(inputs['q_start_loc'], inputs['seq_length']),
-                )
-                return output['logits']
-
-        else:
-            with torch.no_grad():
-                self.tp_model_in_que.put((inputs, swap_in_map, swap_out_map))
-
-            ret_code, output = self.tp_model_out_que.get()
-            if ret_code != 0:
-                logger.error('tp forward failed.')
-                exit()
-
-            return output
+        return self.model_agent.forward(inputs,
+                                        swap_in_map=swap_in_map,
+                                        swap_out_map=swap_out_map)
 
     def step(self, return_logits=False):
         """one step inference. Used to perform streaming chat.
@@ -970,17 +614,19 @@ class Engine:
 
         while True:
             # get all requests
-            num_requests = self.requests.qsize()
 
-            if num_requests == 0 and not self.scheduler.has_unfinished():
-                time.sleep(1)
-                continue
+            with self.mutex:
+                num_requests = self.requests.qsize()
 
-            reqs: List[Request] = []
-            tmp = num_requests
-            while tmp:
-                tmp -= 1
-                reqs.append(self.requests.get())
+                if num_requests == 0 and not self.scheduler.has_unfinished():
+                    time.sleep(1)
+                    continue
+
+                reqs: List[Request] = []
+                tmp = num_requests
+                while tmp:
+                    tmp -= 1
+                    reqs.append(self.requests.get())
 
             # gather requests
             reqs_by_type: Dict[RequestType, Request] = dict(
@@ -1038,6 +684,8 @@ class Engine:
                 else:
                     msg = next(iter(sess.sequences.values()))
                     msg.update_token_ids(req.data['token_ids'])
+                    msg.remain_output_len = req.data['max_request_output_len']
+                    msg.sampling_param = req.data['sampling_param']
                     msg.status = MessageStatus.WAITING
                     self.scheduler.update()
